expr eval: squash warnings around 'srand'
GVPR documents this call as taking and returning `int` values. So going back and
forth between `long`, `long long`, and `time_t` should be fine as these all are
guaranteed to be at least as wide as `int`.

+++ lib/expr/exeval.c
@@ -1155,9 +1155,9 @@ static Extype_t eval(Expr_t *ex, Exnode_t *exnode, void *env) {
 	case SRAND:
 		v.integer = seed;
 		if (exnode->binary) {
-			seed = eval(ex, x, env).integer;
+			seed = (long)eval(ex, x, env).integer;
 		} else
-			seed = time(0);
+			seed = (long)time(0);
 		srand48(seed);
 		return v;
 	case RAND:
